+++ src/Pim/Bundle/CatalogBundle/Entity/Repository/AssociationTypeRepository.php
@@ -25,10 +25,10 @@ class AssociationTypeRepository extends ReferableEntityRepository implements Dat
     {
         $qb = $this->createQueryBuilder('a');
 
-        if ($productAssociations = $product->getProductAssociations()) {
-            $associationTypeIds = $productAssociations->map(
-                function ($productAssociation) {
-                    return $productAssociation->getAssociationType()->getId();
+        if ($associations = $product->getAssociations()) {
+            $associationTypeIds = $associations->map(
+                function ($association) {
+                    return $association->getAssociationType()->getId();
                 }
             );